requirements.txt
absl-py==0.11.0
alabaster==0.7.12
appdirs==1.4.4
astor==0.8.1
astropy==4.2
astroquery==0.4.1
astunparse==1.6.3
attrs==20.3.0
autograd==1.3
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bokeh==2.2.3
boto3==1.16.59
botocore==1.19.59
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
docutils==0.16
eleanor==2.0.1
fbpca==1.0
filelock==3.0.12
flamegraph @ git+https://github.com/evanhempel/python-flamegraph.git@6b70f9068cb987a2b1942fbf048fc88a7a644c40
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gprof2dot==2019.11.30
grpcio==1.32.0
h5py==2.10.0
html5lib==1.1
idna==2.10
imagesize==1.2.0
importlib-metadata==3.4.0
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.3.1
lightkurve @ git+https://github.com/KeplerGO/lightkurve.git@81af7630789ce9bb9fe3ece976091902292327b1
llvmlite==0.35.0
lpproj==0.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
memoization==0.3.2
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
oktopus==0.1.2
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.1
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
photutils==1.0.2
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyerfa==1.7.1.1
Pygments==2.7.4
pyparsing==2.4.7
pytest==6.2.2
pytest-profiling==1.7.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.0
SecretStorage==3.3.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.1
Sphinx==3.4.3
sphinx-automodapi==0.13
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==1.15.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
tess-point==0.6.1
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1
tornado==6.1
tox==3.21.2
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uncertainties==3.1.5
urllib3==1.26.2
virtualenv==20.4.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
